Move AU BECS bank data out of spec
Committed-By-Agent: codex
